le user id plus la vérif

+++ src/controllers/PetController.ts
@@ -8,17 +8,12 @@ class PetController{
 
     static async createPet(req: Request, res: Response) {
       try {
-        const validpet = petSchema.parse(req.body);
-        const newPet = await PetService.createPet(validpet);
+        
+        const newPet = await PetService.createPet(req.body,req.user?.id );
         return ApiResponse.created(res,"Animal créé avec succès",newPet);
 
       } catch (error) {
-        console.error(error);
-        if (error instanceof ZodError) {
-          return ApiResponse.badRequest(res,"Données incorrectes")  
-        }else{
             return ApiResponse.internalServerError(res,"Erreur lors de la crétion de l'animal");
-        }
       }
     }
 
@@ -32,7 +27,7 @@ class PetController{
         const {pets,total} = await PetService.getAllPets({page,perPage});
         const pagination= ApiResponse.createPagination( total, page, perPage);
 
-        return ApiResponse.ok(res,"données récupérées", pets);
+        return ApiResponse.ok(res,"données récupérées", pets,pagination);
 
       } catch (error) {
         return ApiResponse.internalServerError(res,"Erreur lors de la récupération des animaux");
@@ -48,7 +43,10 @@ class PetController{
           return ApiResponse.badRequest(res, "ID de l'animal requis");
         }
 
-        const pet = await PetService.getPetById(id);
+        const pet = await PetService.getPetById(id)
+        if (pet?.user_id !== req.user?.id){
+          return ApiResponse.badRequest(res,"pas ton pet ")
+        }
 
         if (!pet) {
           return ApiResponse.notFound(res,"Animal non trouvé");
@@ -62,6 +60,10 @@ class PetController{
     static async updatePet(req: Request, res: Response) {
       try {
         const { id } = req.params;
+        const pet = await PetService.getPetById(id)
+        if (pet?.user_id !== req.user?.id){
+          return ApiResponse.badRequest(res,"pas ton pet ")
+        }
         const updatedPet = await PetService.updatePet(id, req.body);
         if (!updatedPet) {
           return ApiResponse.notFound(res,"Animal non trouvé");
@@ -75,6 +77,16 @@ class PetController{
     static async deletePet(req: Request, res: Response) {
       try {
         const { id } = req.params;
+
+        if (!id) {
+          return ApiResponse.badRequest(res, "ID de l'animal requis");
+        }
+
+        const pet = await PetService.getPetById(id)
+        if (pet?.user_id !== req.user?.id){
+          return ApiResponse.badRequest(res,"pas ton pet")
+        }
+
         const deleted = await PetService.deletePet(id);
         if (!deleted) {
           return ApiResponse.notFound(res,"Animal non trouvé");
